Create Contract (Library)
This is a smart contract deployed on flow blockchain using cadence programming language. It used to manage a digital library system.
it contains an array used to store books.it also has a dictionary with two properties "title" and "shelf number". It has two functions to help modify the contract.